CommuteCA: an R package to develop standardized methods for transport analysis in research

The main objective of the CommuteCA package is to develop standardized methods for accessibility analysis in research, especially for studies using Statistics Canada surveys. Among the available surveys, we focused our efforts on the 2021 Census of Population, which contain valuable variables for transportation research. This package was created in conjunction with the office of the Research Data Center at McMaster University, the Sherman Centre for Digital Scholarship and the Mobilizing Justice.

References

https://github.com/dias-bruno/CommuteCA

See also

Author

Bruno Dos Santos, School of Earth, Environment and Society, McMaster University dossanb@mcmaster.ca

Antonio Paez, School of Earth, Environment and Society, McMaster University paezha@mcmaster.ca